% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_list_species_range}
\alias{BIEN_list_species_range}
\title{Extract a list of all species in the BIEN database.}
\usage{
BIEN_list_species_range(country, elevation_min, elevation_max, ...)
}
\arguments{
\item{country}{A single country or a vector of country.}

\item{elevation_min}{Low elevation}

\item{elevation_max}{Max elevation}

\item{...}{Additional arguments passed to internal functions.}
}
\value{
Dataframe containing a list of all species in the BIEN database.
}
\description{
BIEN_list_all produces a list of all species in the BIEN database.
}
\examples{
\dontrun{
species_list<-BIEN_list_species()}
}
\concept{list functions}
\keyword{internal}
