% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN.R
\name{BIEN_metadata_list_political_names}
\alias{BIEN_metadata_list_political_names}
\title{List political divisions and associated geonames codes.}
\usage{
BIEN_metadata_list_political_names(...)
}
\arguments{
\item{...}{Additional arguments passed to internal functions.}
}
\value{
A dataframe containing political division names and their associated codes.
}
\description{
BIEN_metadata_list_political_names downloads country, state, and county names and associated codes used by BIEN.
}
\note{
Political names and codes follow http://www.geonames.org/
}
\examples{
\dontrun{
BIEN_metadata_list_political_names()}
}
\concept{metadata functions}
\keyword{internal}
