% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_readme_rmd.R
\name{create_readme_rmd}
\alias{create_readme_rmd}
\title{Create a README.Rmd file in a specified path}
\usage{
create_readme_rmd(path)
}
\arguments{
\item{path}{A character string specifying the path where the README.Rmd file will be created.}
}
\value{
None
}
\description{
This function creates a new folder if it does not exist and creates a README.Rmd file with default content in the specified path.
}
\examples{
#create_readme_rmd("~/my_project")
}
