% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_auxiliar.R
\name{.classify_algo}
\alias{.classify_algo}
\title{Clasificador interno de nombres científicos por categorías infraespecíficas}
\usage{
.classify_algo(x_split_i, Infrasp_cat_reg)
}
\arguments{
\item{x_split_i}{Un vector de caracteres que representa el nombre científico
separado por espacios (por ejemplo, \code{strsplit(nombre, " ")[[1]]}).}

\item{Infrasp_cat_reg}{Expresiones regulares que representan abreviaturas
comunes de categorías infraespecíficas (por ejemplo, "SUBSP.", "VAR.", "F.").}
}
\value{
Un vector de 10 elementos que incluye:
\describe{
\item{Elemento 1}{Género}
\item{Elemento 2}{Epíteto específico}
\item{Elemento 3}{Nombre del autor (si corresponde)}
\item{Elemento 4}{(Reservado, sin uso en esta función)}
\item{Elemento 5}{Subespecie}
\item{Elemento 6}{(Reservado, eliminado por funciones posteriores)}
\item{Elemento 7}{Variedad}
\item{Elemento 8}{Subvariedad}
\item{Elemento 9}{Forma}
\item{Elemento 10}{Subforma}
}
}
\description{
Esta función interna implementa el algoritmo que clasifica un nombre
científico estandarizado en sus componentes taxonómicos: género, epíteto específico,
autoría y posibles categorías infraespecíficas (subespecie, variedad, forma, etc.).
Se utiliza como parte del flujo de trabajo de estandarización y descomposición
de nombres en análisis taxonómicos.
}
\details{
Esta función busca patrones específicos de abreviaturas infraespecíficas dentro
del nombre. Si encuentra alguna, asigna el nombre correspondiente a la posición
de salida apropiada. Además, identifica y asigna el nombre del autor si no forma
parte del nombre infraespecífico.
}
\keyword{internal}
