% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_auxiliar.R
\name{.names_standardize}
\alias{.names_standardize}
\title{Standardize taxonomic names}
\usage{
.names_standardize(splist)
}
\arguments{
\item{splist}{A character vector containing scientific names to be standardized.}
}
\value{
A character vector of standardized names. Transformations include:
\itemize{
\item Conversion to uppercase.
\item Removal of prefixes such as \code{CF.} and \code{AFF.}.
\item Trimming of leading and trailing whitespace.
\item Replacing underscores (\verb{_}) with spaces.
\item Standardization of infraspecific ranks (e.g. \code{VAR}, \code{F.}, \code{SUBSP.}).
\item Removal of hybrid indicators (\code{X} or \verb{×}).
\item Elimination of non-alphabetical symbols at the beginning of the name.
}
}
\description{
This internal function standardizes scientific names by applying a set of
transformations to improve consistency before further parsing or matching.
}
\details{
Hybrid symbols (e.g., \code{"X"} or \code{"×"}) are removed with a warning listing
the affected names.
}
\examples{
splist <- c("cf. Abies alba", "Pinus_sylvestris", "X Quercus robur")
.names_standardize(splist)

}
\keyword{internal}
