% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splist_auxiliar.R
\name{.transform_split_classify}
\alias{.transform_split_classify}
\title{Transform classified species data}
\usage{
.transform_split_classify(df)
}
\arguments{
\item{df}{Data frame generado por la función \code{.splist_classify()}, el cual contiene las columnas
\code{orig_genus}, \code{orig_species}, \code{subspecies}, \code{variety}, \code{subvariety}, \code{forma}, \code{subforma}.}
}
\value{
Un data frame con columnas adicionales:
\describe{
\item{sorter}{Índice para mantener el orden original.}
\item{orig_infraspecies}{Nombre de la infraspecie, si aplica.}
\item{infra_rank}{Categoría taxonómica de la infraspecie (ej. SUBSP., VAR., etc.).}
\item{rank}{Nivel taxonómico estimado: 1 = género, 2 = especie, 3 = infraspecie.}
}
}
\description{
Esta función transforma una tabla resultante de la clasificación taxonómica de nombres científicos,
añadiendo columnas auxiliares como \code{orig_infraspecies}, \code{infra_rank} y \code{rank}, útiles para análisis posteriores.
}
\keyword{internal}
