% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pesa_caption.R
\name{get_social_icons}
\alias{get_social_icons}
\title{Get Social Media Icons}
\usage{
get_social_icons()
}
\value{
A named list containing the HTML code for each social media icon (LinkedIn, GitHub, Twitter, and Bluesky).
}
\description{
Provides a list of HTML strings representing social media icons using Font Awesome 6.
}
\examples{
icons <- get_social_icons()
icons$twitter # Returns the HTML for the Twitter icon
}
