% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_commit.R
\name{gitcommit}
\alias{gitcommit}
\title{Git Commit}
\usage{
gitcommit(msg = "commit from Rstudio", dir = getwd())
}
\arguments{
\item{msg}{message}

\item{dir}{directory name}
}
\value{
a list of commited files
}
\description{
Git Commit
}
