% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_form_uicn.R
\name{growth_forms_iucn}
\alias{growth_forms_iucn}
\title{Get the IUCN growth form of a species}
\usage{
growth_forms_iucn(sp, key = NULL)
}
\arguments{
\item{sp}{A character vector of species names}

\item{key}{A character string for the IUCN API key (optional)}
}
\value{
A tibble with species, species_2, and growth_form
}
\description{
This function takes a species name and returns the IUCN growth form
}
\examples{
\dontrun{
growth_forms_iucn("Gorilla gorilla")
}

}
