% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iucn_table.R
\name{iucn_table}
\alias{iucn_table}
\title{IUCN red list}
\usage{
iucn_table(x, key = NULL)
}
\arguments{
\item{x}{Vector of species names}

\item{key}{IUCN api key}
}
\value{
Tibble
}
\description{
IUCN red list
}
\examples{
#iucn_table("Schinus molle")
}
