% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ij_loop.R
\name{run_ij_loop}
\alias{run_ij_loop}
\title{Automated leaf area analysis}
\usage{
run_ij_loop(path, new_path, output, ...)
}
\arguments{
\item{path}{Set directory that contains leaf images.}

\item{new_path}{Set temp directory that contains leaf images.}

\item{output}{Set output directory that contains tiff images.}

\item{...}{}
}
\value{
A tibble.
}
\description{
Automated leaf area analysis
}
\examples{
#dir <- paste0(here::here(), "/", "inst/images")
#temp <-  paste0(here::here(), "/", "inst/temp")
#output <-  paste0(here::here(), "/", "inst/output")
#data <- run_ij_p(dir, temp, output)
#data
}
