% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_write.R
\name{shape_write}
\alias{shape_write}
\title{Write Simple Features to an ESRI Shapefile}
\usage{
shape_write(sf, folder, filename)
}
\arguments{
\item{sf}{A Simple Features (sf) object to be written to the Shapefile.}

\item{folder}{The folder where the Shapefile will be created.}

\item{filename}{The desired filename for the Shapefile (without the .shp extension).}
}
\value{
None. The function writes the SF object to a Shapefile.
}
\description{
This function writes a Simple Features (sf) object to an ESRI Shapefile.
It drops the Z and M dimensions from the object before writing.
}
\examples{
\dontrun{
# Write an SF object to a Shapefile
shape_write(my_sf_object, "output_folder", "output_filename")
}

}
