% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalize a string and remove unnecessary white spaces}
\usage{
simple_cap(x)
}
\arguments{
\item{x}{The input string to be capitalized and cleaned from spaces.}
}
\value{
A string with the first letter capitalized and unnecessary white spaces removed.
}
\description{
This function takes a string as input and performs the following operations:
\enumerate{
\item Removes unnecessary white spaces between words.
\item Capitalizes the first letter of the string and converts the rest to lowercase.
}
}
\examples{
\dontrun{
# Example usage
simple_cap("Miconia    secundifolia  subsp.  Secundifolia")
# Output: "Miconia secundifolia subsp. secundifolia"
}

}
