% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_and_assing.R
\name{split_and_assign}
\alias{split_and_assign}
\title{Divide un dataframe en subgrupos y asigna cada subgrupo a un objeto con un
nombre limpio}
\usage{
split_and_assign(df, variable)
}
\arguments{
\item{df}{El dataframe que se dividirá en subgrupos.}

\item{variable}{La variable por la cual se dividirá el dataframe en subgrupos.}
}
\value{
No devuelve nada explícitamente, pero asigna los subgrupos a objetos
con nombres limpios en el entorno global.
}
\description{
Esta función toma un dataframe y una variable por la cual dividirlo en
subgrupos. Luego, asigna cada subgrupo a un objeto con un nombre limpio
generado a partir de los valores únicos de la variable.
}
\examples{
\dontrun{
# Ejemplo de uso con iris
split_and_assign(iris, "Species")
}
}
