% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_delete_after.R
\name{str_delete_after}
\alias{str_delete_after}
\title{Delete characters in a string which occur after a specified pattern}
\usage{
str_delete_after(string, pattern, pos = 1)
}
\arguments{
\item{string}{The string from which to extract.}

\item{pattern}{Pattern to look for. The default interpretation is a regular expression.}

\item{pos}{Position of the character used as a marker.}
}
\value{
The part of the string before the pattern.
}
\description{
Vectorised over 'string' and 'pattern'.
}
