% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrs.R
\name{tnrs}
\alias{tnrs}
\title{tnrs}
\usage{
tnrs(df, specie)
}
\arguments{
\item{df}{A data.frame}

\item{specie}{Species column name}
}
\value{
a tibble
}
\description{
tnrs
}
\examples{
#df \%>\% tnrs(specie)
}
