% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnrs_vec.R
\name{tnrs_vec}
\alias{tnrs_vec}
\title{TNRS vectorized}
\usage{
tnrs_vec(x)
}
\arguments{
\item{x}{species names vector}
}
\value{
a tibble
}
\description{
TNRS vectorized
}
\examples{
x <- c("Plantago tubuloso", "Plantago tubulosa")
tnrs_vec(x)
}
