% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_na.R
\name{check_na}
\alias{check_na}
\title{Count all missing values in a tibble}
\usage{
check_na(df)
}
\arguments{
\item{df}{A tibble or data.frame.}
}
\value{
A tibble displaying the number of missing values in the dataframe.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
